Extract config from module
Since modules no longer have to be jammed into the same package, clutter from a single small module is less of a concern.
